refactor: 도메인명을 bookclub -> club으로 변경
