--- bookduck/src/main/java/com/mmc/bookduck/domain/bookclub/entity/ClubStatus.java
@@ -1,5 +1,0 @@
-package com.mmc.bookduck.domain.bookclub.entity;
-
-public enum ClubStatus {
-    ACTIVE, CLOSED
-}
