+++ bookduck/src/main/java/com/mmc/bookduck/domain/club/dto/response/ClubSearchResponseDto.java
@@ -1,7 +1,7 @@
-package com.mmc.bookduck.domain.bookclub.dto.response;
+package com.mmc.bookduck.domain.club.dto.response;
 
-import com.mmc.bookduck.domain.bookclub.entity.Club;
-import com.mmc.bookduck.domain.bookclub.entity.ClubStatus;
+import com.mmc.bookduck.domain.club.entity.Club;
+import com.mmc.bookduck.domain.club.entity.ClubStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
