+++ bookduck/src/main/java/com/mmc/bookduck/domain/club/dto/response/ClubSummaryResponseDto.java
@@ -1,6 +1,6 @@
-package com.mmc.bookduck.domain.bookclub.dto.response;
+package com.mmc.bookduck.domain.club.dto.response;
 
-import com.mmc.bookduck.domain.bookclub.entity.Club;
+import com.mmc.bookduck.domain.club.entity.Club;
 import lombok.Builder;
 
 @Builder
