+++ bookduck/src/main/java/com/mmc/bookduck/domain/club/entity/ClubStatus.java
@@ -1,0 +1,5 @@
+package com.mmc.bookduck.domain.club.entity;
+
+public enum ClubStatus {
+    ACTIVE, CLOSED
+}
