+++ bookduck/src/main/java/com/mmc/bookduck/domain/club/repository/ClubMemberReadStatusRepository.java
@@ -1,7 +1,7 @@
-package com.mmc.bookduck.domain.bookclub.repository;
+package com.mmc.bookduck.domain.club.repository;
 
-import com.mmc.bookduck.domain.bookclub.entity.ClubMember;
-import com.mmc.bookduck.domain.bookclub.entity.ClubMemberReadStatus;
+import com.mmc.bookduck.domain.club.entity.ClubMember;
+import com.mmc.bookduck.domain.club.entity.ClubMemberReadStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
